app/celery_app.py
from celery import Celery
import os
from dotenv import load_dotenv

# Загрузка переменных окружения из .env файла
load_dotenv()

# Создание экземпляра Celery
celery_app = Celery('car_advisor')

# Настройка конфигурации Celery
celery_app.conf.update(
    broker_url=os.getenv('CELERY_BROKER_URL', 'redis://redis:6379/0'),
    result_backend=os.getenv('CELERY_RESULT_BACKEND', 'redis://redis:6379/0'),
    task_serializer='json',
    accept_content=['json'],
    result_serializer='json',
    timezone='UTC',
    enable_utc=True,
)

# Автоматическое обнаружение задач
celery_app.autodiscover_tasks(['app.tasks'])