requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
redis==5.0.1
celery==5.3.4
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
requests==2.31.0
psycopg2-binary==2.9.9
cachetools==5.3.2
loguru==0.7.2
faker==20.1.0